requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.2
dlib==20.0.0
exceptiongroup==1.3.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
python-multipart==0.0.20
requests==2.32.4
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
